requirements.txt
fastapi==0.101.1
uvicorn==0.22.0
assemblyai==0.41.1
python-dotenv==1.0.0
python-multipart==0.0.20
google-cloud-speech==2.21.0
google-cloud-storage==3.1.0
openai==1.54.4
